refactor: Phase 1 - Modularize Eleventy configuration architecture
BREAKING CHANGE: Restructured .eleventy.js from monolithic to modular architecture

## Summary
- Refactored 448-line monolithic .eleventy.js into 178-line orchestration file
- Extracted configuration into 5 specialized modules in eleventy/config/
- Integrated sitemap absolute URL fixes to resolve link-checker issues
- Maintained 100% functional compatibility with improved maintainability

## Architecture Changes
### Main Configuration (.eleventy.js)
- Reduced from 448 to 178 lines (60% reduction)
- Added modular imports and environment detection
- Integrated siteDomain global data for absolute URLs
- Production-first domain detection (https://a11y.canada.ca)

### New Configuration Modules
- eleventy/config/markdown.js - Markdown-it setup with plugins
- eleventy/config/filters.js - 14 template filters (dates, formatting, utils)
- eleventy/config/transforms.js - 3 HTML transforms (minify, sa11y, debug)
- eleventy/config/collections.js - 4 content collections with git integration
- eleventy/config/passthrough.js - Static file copying configuration

## Technical Improvements
- **Maintainability**: Separated concerns for easier debugging and collaboration
- **Readability**: Each module focuses on single responsibility
- **Scalability**: Foundation for Phase 2 performance optimizations
- **Link-checker fix**: Sitemap now generates absolute URLs correctly

## Performance Impact
- Build time: ~1-3 second overhead (acceptable for maintainability gains)
- Same output: 257 files generated consistently
- Module loading overhead offset by future optimization opportunities

## Testing & Validation
- ✅ Successful builds on both main and refactor branches
- ✅ All 257 files generated correctly
- ✅ Automated testing infrastructure implemented
- ✅ Link-checker functionality restored

## Next Steps
- Phase 2: Performance optimizations (memoization, conditional loading)
- Phase 3: Advanced modularization (plugin architecture)
