+++ .eleventy.js
@@ -1,281 +1,61 @@
+/**
+ * Eleventy Configuration - Main Entry Point
+ *
+ * Modular configuration for the Government of Canada Digital Accessibility Toolkit.
+ * This file orchestrates all configuration modules for better maintainability.
+ *
+ * Refactored: Phase 1 - Extract Configuration Objects
+ * Date: September 26, 2025
+ */
+
 const fs = require('fs');
 const path = require('path');
 const { execSync } = require('child_process');
-const nunjucks = require('nunjucks');  // Import Nunjucks for rendering
-const markdownIt = require('markdown-it');
-const markdownItAnchor = require('markdown-it-anchor');
-const markdownItAttrs = require('markdown-it-attrs');
 const { EleventyHtmlBasePlugin } = require('@11ty/eleventy');
-const { stripHtml } = require('string-strip-html');
-const beautify = require('js-beautify').html;
-const { DateTime } = require('luxon');
 const { getChangedPagesForBuild } = require('./scripts/build-changed-pages');
-const changedPageUrls = [];
 
-// ANSI escape codes for blue underline
-const underline = '\x1b[4m';  // Underline font
-const resetColor = '\x1b[0m'; // Reset font to default
+// Import configuration modules
+const configureMarkdown = require('./eleventy/config/markdown');
+const configureFilters = require('./eleventy/config/filters');
+const configureTransforms = require('./eleventy/config/transforms');
+const configureCollections = require('./eleventy/config/collections');
+const configurePassthrough = require('./eleventy/config/passthrough');
+
+// ANSI escape codes for colored console output
+const underline = '\x1b[4m';
+const resetColor = '\x1b[0m';
 
 // Check if we are in watch mode (development)
 const isWatchMode = process.env.ELEVENTY_WATCH === 'true';
 
+// Global variables for change tracking (to be improved in Phase 2)
+global.changedFilePaths = new Set();
+global.gitChangedUrls = [];
+
 module.exports = function (eleventyConfig) {
-  const eleventySlugify = eleventyConfig.getFilter('slug');
 
-  let markdownItOptions = {
-    html: true, // Enable HTML tags in Markdown
-  };
+  // Configure markdown processing and get markdown instance
+  const markdownInstance = configureMarkdown(eleventyConfig);
 
-  const md = markdownIt(markdownItOptions)
-    .use(markdownItAttrs)
-    .use(markdownItAnchor, {
-      slugify: s => eleventySlugify(s, { lower: true, strict: true, locale: 'fr' }),
-      permalink: false, // Disable permalinks
-    });
+  // Configure all other modules
+  configureFilters(eleventyConfig, markdownInstance);
+  configureTransforms(eleventyConfig);
+  configureCollections(eleventyConfig, markdownInstance);
+  configurePassthrough(eleventyConfig);
 
-  eleventyConfig.setLibrary("md", md);
-
-  eleventyConfig.addShortcode('extractHeadings', function (content, tocType) {
-    const tokens = md.parse(content, {});
-    const levels = tocType === 'tocSimple' ? 1 : 2; // tocSimple only includes level 2 headings
-    const validTags = Array.from({ length: levels + 1 }, (_, i) => `h${i + 2}`);
-    const headings = tokens.filter(token =>
-      validTags.includes(token.tag) && token.type === 'heading_open'
-    ).map(token => {
-      const level = token.tag;
-      const rawText = tokens[tokens.indexOf(token) + 1].content;
-      const text = stripHtml(rawText).result; // Strip HTML tags from the heading text
-      const id = eleventySlugify(text, { lower: true, strict: true, locale: 'fr' });
-      return { level, text, id };
-    });
-
-    // Create TOC HTML
-    let tocHTML = '<aside><h2>{{ onThisPage[locale].heading }}</h2><ul>';
-    const levelsStack = [];
-
-    headings.forEach(heading => {
-      const levelIndex = parseInt(heading.level.substring(1)) - 1;
-
-      while (levelsStack.length && levelsStack[levelsStack.length - 1] > levelIndex) {
-        tocHTML += '</ul></li>';
-        levelsStack.pop();
-      }
-
-      if (levelsStack.length && levelsStack[levelsStack.length - 1] === levelIndex) {
-        tocHTML += '</li>';
-      }
-
-      if (!levelsStack.length || levelsStack[levelsStack.length - 1] < levelIndex) {
-        tocHTML += '<ul>';
-        levelsStack.push(levelIndex);
-      }
-
-      tocHTML += `<li><a href="#${heading.id}">${heading.text}</a>`;
-    });
-
-    while (levelsStack.length) {
-      tocHTML += '</ul></li>';
-      levelsStack.pop();
-    }
-
-    tocHTML += '</ul></aside>';
-
-    return tocHTML;
-  });
-
-  eleventyConfig.addFilter("stripTagsSlugify", (str) => {
-    if (!str) return;
-    return eleventySlugify(stripHtml(str).result, { lower: true, strict: true, locale: 'fr' });
-  });
-
+  // Add core Eleventy plugins
   eleventyConfig.addPlugin(EleventyHtmlBasePlugin);
 
-  // Format HTML output
-  eleventyConfig.addTransform("htmlbeautify", function (content, outputPath) {
-    if (outputPath && outputPath.endsWith(".html")) {
-      return beautify(content, {
-        indent_size: 2,
-        indent_char: ' ',
-        preserve_newlines: false,
-        max_preserve_newlines: 1,
-        wrap_line_length: 0,
-        end_with_newline: true,
-      });
-    }
-    return content;
-  });
-
-  // Add UTF-8 BOM to CSV files for proper Excel encoding
-  eleventyConfig.addTransform("csvUtf8Bom", function (content, outputPath) {
-    if (outputPath && outputPath.endsWith(".csv")) {
-      // For Excel compatibility, we need to ensure the content is properly encoded
-      // Try a different approach with UTF-8 BOM and ensure content is UTF-8
-      const utf8Bom = '\uFEFF'; // Use Unicode BOM character instead of bytes
-      return utf8Bom + content;
-    }
-    return content;
-  });
-
-  eleventyConfig.addFilter("localeMatch", function (collection) {
-    const { locale } = this.ctx; // avoid retrieving it for each item
-    return collection.filter((item) => item.data.locale === locale);
-  });
-
-  // Specific postDate filter for blog posts or other content with strict formatting
-  eleventyConfig.addFilter("postDate", (dateObj) => {
-    return DateTime.fromJSDate(dateObj, { zone: "utc" })
-      .setLocale("en")
-      .toFormat("yyyy'-'MM'-'dd");
-  });
-
-  // Format date filter to display full date
-  eleventyConfig.addFilter("formatDate", function (dateObj) {
-    const locale = this.ctx.locale || 'en';  // Use the locale from the context, default to 'en'
-
-    let luxonDate;
-
-    // Check if the dateObj is an ISO string and parse it to a DateTime object
-    if (typeof dateObj === 'string') {
-      luxonDate = DateTime.fromISO(dateObj);  // Handle ISO strings
-    } else if (dateObj instanceof Date) {
-      luxonDate = DateTime.fromJSDate(dateObj);  // Handle JS Date objects
-    } else {
-      return "Invalid Date";  // Return fallback if neither works
-    }
-
-    return luxonDate.setLocale(locale).toLocaleString(DateTime.DATE_HUGE);  // Use DATE_HUGE for formatting
-  });
-
-  // Format year and month to display "Month Year" in the correct locale
-  eleventyConfig.addFilter("formatYearMonth", function (dateString) {
-    const locale = this.ctx.locale || 'en';  // Use the locale from the context, default to 'en'
-
-    return DateTime.fromFormat(dateString, 'yyyy-MM')
-      .setLocale(locale)
-      .toFormat('LLLL yyyy');  // Use LLLL for full month name
-  });
-
-  eleventyConfig.addFilter("percentage", function (value) {
-    return (parseFloat(value) * 100).toFixed(2) + '%';
-  });
-
-  eleventyConfig.addFilter("recentMonths", function (analytics) {
-    const sortedMonths = Object.entries(analytics).sort(([a], [b]) => b.localeCompare(a)); // Reverse sort by month key
-    return sortedMonths.slice(0, 3); // Get the last three months
-  });
-
-  eleventyConfig.addPassthroughCopy({ "./src/_docs": "docs" });
-  eleventyConfig.addPassthroughCopy({ "./src/_images": "img" });
-  eleventyConfig.addPassthroughCopy({ "./src/CNAME": "CNAME" });
-  eleventyConfig.addPassthroughCopy({ "./src/_scripts": "js" });
-
-  eleventyConfig.addCollection("allHeadings", function (collectionApi) {
-    return collectionApi.getAll().map(item => {
-      if (item.data.toc || item.data.tocSimple) {
-        const tokens = md.parse(item.template.frontMatter.content, {});
-        const levels = item.data.tocSimple ? 1 : 2; // tocSimple only includes level 2 headings
-        const validTags = Array.from({ length: levels + 1 }, (_, i) => `h${i + 2}`);
-        const headings = tokens.filter(token =>
-          validTags.includes(token.tag) && token.type === 'heading_open'
-        ).map(token => {
-          const level = token.tag;
-          const rawText = tokens[tokens.indexOf(token) + 1].content;
-          const text = stripHtml(rawText).result; // Strip HTML tags from the heading text
-          const id = eleventySlugify(text, { lower: true, strict: true, locale: 'fr' });
-          return { level, text, id };
-        });
-        item.data.headings = headings;
-      }
-      return item;
-    });
-  });
-
-  // Custom collection for changed pages
-  eleventyConfig.addCollection('changedPages', function (collectionApi) {
-    const changedPages = [];
-    collectionApi.getAll().forEach(item => {
-      const normalizedInputPath = path.relative('./', item.inputPath);
-      if (changedFilePaths.has(normalizedInputPath)) {
-        // Store both the URL, title, and locale for each changed page
-        changedPages.push({
-          url: item.url,
-          title: item.data.title || item.fileSlug, // fallback to fileSlug if no title
-          locale: item.data.locale || 'en' // default to 'en' if no locale is set
-        });
-      }
-    });
-    return changedPages; // Returning the changed URLs, titles, and locales for the template
-  });
-
-  // Get the role keys that belong to a given group
-  eleventyConfig.addFilter("roleKeysForGroup", function (groupKey, rolesData) {
-    if (!groupKey || !rolesData || !rolesData.roles) return [];
-    return Object.entries(rolesData.roles)
-      .filter(([, meta]) => meta.group === groupKey)
-      .map(([key]) => key);
-  });
-
-  // Given a list of role keys, return all pages that match ANY of them
-  eleventyConfig.addFilter("byAnyRole", function (collection, roleKeys) {
-    if (!collection || !roleKeys || !roleKeys.length) return [];
-    return collection.filter((item) => {
-      const r = item.data.role;
-      if (!r) return false;
-      const arr = Array.isArray(r) ? r : [r];
-      return arr.some((k) => roleKeys.includes(k));
-    });
-  });
-
-  // Collection of all pages that have at least one role
-  eleventyConfig.addCollection("rolePages", (api) =>
-    api.getAll().filter((item) => !!item.data.role)
-  );
-
-  // Custom collection for role groups to work around pagination tag issues
-  eleventyConfig.addCollection("roleGroup", (api) => {
-    // Find all pages generated from the roles-group.njk template OR
-    // pages with URLs matching role group patterns
-    return api.getAll().filter((item) => {
-      // Check if the page was generated from the roles-group template
-      const fromRoleGroupTemplate = item.inputPath && item.inputPath.includes('roles-group.njk');
-
-      // Check if URL matches role group pattern (e.g., /en/roles/business/, /fr/roles/design/)
-      const matchesRoleGroupPattern = item.url && item.url.match(/^\/(en|fr)\/roles\/(administration|author|business|design|development|testing)\/$/);
-
-      return fromRoleGroupTemplate || matchesRoleGroupPattern;
-    });
-  });
-
-  // Add custom Markdown filter for Nunjucks
-  eleventyConfig.addNunjucksFilter("markdown", function (value) {
-    return md.render(value);
-  });
-
-  // Add split filter for Nunjucks
-  eleventyConfig.addNunjucksFilter("split", function (str, separator) {
-    if (typeof str !== 'string') return [];
-    return str.split(separator || ' ');
-  });
-
-  // Add wordCount filter for convenience
-  eleventyConfig.addNunjucksFilter("wordCount", function (content) {
-    if (!content) return 0;
-    const text = typeof content === 'string' ? content : String(content);
-    const stripped = stripHtml(text).result;
-    const words = stripped.trim().split(/\s+/);
-    return words.length === 1 && words[0] === '' ? 0 : words.length;
-  });
-
-  let changedFilesMap = new Map();
-  let changedFilePaths = new Set(); // Track changed file paths for collections and transforms
+  // Changed pages tracking system
   let changedPages = [];
-  let gitChangedUrls = [];
+  let changedFilesMap = new Map();
+
+  // Environment detection for URL generation
   let domain = 'https://a11y.canada.ca'; // Default to production
   let port = '8080'; // Default port
 
   // Detect GitHub Codespaces
   if (process.env.CODESPACES && process.env.CODESPACE_NAME) {
-    // Use the correct format for Codespaces URLs with the .app domain
     domain = `https://${process.env.CODESPACE_NAME}-${port}.app.github.dev`;
   }
 
@@ -290,14 +70,16 @@ module.exports = function (eleventyConfig) {
     domain = `http://localhost:${port}`;
   }
 
-  // Add domain as global data for templates
+  // Add domain as global data for templates (used by sitemap)
   eleventyConfig.addGlobalData("siteDomain", domain);
+
+  // Event Handlers
 
   // Capture changed files before the build starts (Method 1: Eleventy watch)
   eleventyConfig.on('beforeWatch', (changedFiles) => {
     changedFiles.forEach(file => {
       console.log(`Changed source file: ${file}`);
-      changedFilesMap.set(file, null); // Track the file
+      changedFilesMap.set(file, null);
     });
   });
 
@@ -309,11 +91,11 @@ module.exports = function (eleventyConfig) {
       skipFetch: isWatchMode  // Skip network fetch in development watch mode
     });
 
-    // Populate changedFilePaths Set for use by collections and transforms
-    changedFilePaths.clear();
+    // Populate global changedFilePaths Set for use by collections and transforms
+    global.changedFilePaths.clear();
     changedPages.forEach(page => {
       if (page.inputPath) {
-        changedFilePaths.add(page.inputPath);
+        global.changedFilePaths.add(page.inputPath);
       }
     });
 
@@ -322,49 +104,16 @@ module.exports = function (eleventyConfig) {
     }
   });
 
-  // Hook into the HTML generation process (logging to the console)
-  eleventyConfig.addTransform("captureGeneratedUrl", function (content, outputPath) {
-    if (outputPath && outputPath.endsWith('.html')) {
-      const inputPath = path.relative('./', this.page.inputPath);
-
-      // Check if this file was changed based on Git diff
-      if (changedFilePaths.has(inputPath)) {
-        // Adjust URL generation to prevent double port addition
-        let fullUrl = domain;
-
-        // Only add the port if it's localhost and it hasn't been added already
-        if (domain.includes('localhost') && !domain.includes(`:${port}`)) {
-          fullUrl += `:${port}`;
-        }
-
-        // Append the page URL
-        fullUrl += this.page.url;
-
-        if (isWatchMode) {
-          // Log individual URLs for each changed page in dev mode
-          console.log(`${underline}Captured URL: ${fullUrl}${resetColor}`);
-        } else {
-          // Log file paths in production mode
-          console.log(`${underline}Changed file: ${inputPath}${resetColor}`);
-        }
-
-        // Track the URL or file path for summary logging later
-        gitChangedUrls.push(isWatchMode ? fullUrl : inputPath);
-      }
-    }
-    return content;
-  });
-
-  // After build, log a summary and provide a link to the review page in local development
+  // After build summary and review page link
   eleventyConfig.on('afterBuild', () => {
-    const changedFilesCount = changedFilePaths.size;
+    const changedFilesCount = global.changedFilePaths.size;
 
     if (changedFilesCount > 0) {
       // Log the summary of changed pages
       console.log(`\n${changedFilesCount} page(s) changed.`);
 
       // Log individual URLs or file paths, depending on mode
-      gitChangedUrls.forEach((changedItem) => {
+      global.gitChangedUrls.forEach((changedItem) => {
         console.log(`${underline}${changedItem}${resetColor}`);
       });
 
@@ -385,12 +134,13 @@ module.exports = function (eleventyConfig) {
       console.log('No pages to review.\n');
     }
 
-    // Clear the set for the next watch cycle
-    changedFilePaths.clear();
-    gitChangedUrls = [];
+    // Clear for the next watch cycle
+    global.changedFilePaths.clear();
+    global.gitChangedUrls = [];
   });
 
   // Add computed data for git creation dates
+  // TODO: This will be optimized with memoization in Phase 2
   eleventyConfig.addGlobalData("eleventyComputed", {
     gitCreated: (data) => {
       if (data.page && data.page.inputPath) {
@@ -403,7 +153,6 @@ module.exports = function (eleventyConfig) {
           }
         } catch (error) {
           // Silently handle errors - some files might not have git history
-          // console.error(`Error getting git creation date for ${data.page.inputPath}:`, error.message);
         }
       }
       return null;
@@ -414,6 +163,7 @@ module.exports = function (eleventyConfig) {
   eleventyConfig.ignores.add("src/pages/_template.md");
   eleventyConfig.ignores.add("src/links/_template.md");
 
+  // Return Eleventy configuration
   return {
     dir: {
       input: "src",
