+++ eleventy/config/markdown.js
@@ -1,0 +1,86 @@
+/**
+ * Eleventy Markdown Configuration
+ *
+ * This module contains markdown processing setup for the Digital Accessibility Toolkit.
+ * Includes markdown-it configuration with plugins for attributes and anchors.
+ */
+
+const markdownIt = require('markdown-it');
+const markdownItAnchor = require('markdown-it-anchor');
+const markdownItAttrs = require('markdown-it-attrs');
+const { stripHtml } = require('string-strip-html');
+
+module.exports = function (eleventyConfig) {
+
+  // Get Eleventy's built-in slugify function
+  const eleventySlugify = eleventyConfig.getFilter('slug');
+
+  // Markdown-it options
+  let markdownItOptions = {
+    html: true, // Enable HTML tags in Markdown
+  };
+
+  // Create and configure markdown-it instance
+  const md = markdownIt(markdownItOptions)
+    .use(markdownItAttrs)
+    .use(markdownItAnchor, {
+      slugify: s => eleventySlugify(s, { lower: true, strict: true, locale: 'fr' }),
+      permalink: false, // Disable permalinks
+    });
+
+  // Set the markdown library for Eleventy
+  eleventyConfig.setLibrary("md", md);
+
+  // Add shortcode for extracting headings and generating table of contents
+  eleventyConfig.addShortcode('extractHeadings', function (content, tocType) {
+    const tokens = md.parse(content, {});
+    const levels = tocType === 'tocSimple' ? 1 : 2; // tocSimple only includes level 2 headings
+    const validTags = Array.from({ length: levels + 1 }, (_, i) => `h${i + 2}`);
+
+    const headings = tokens.filter(token =>
+      validTags.includes(token.tag) && token.type === 'heading_open'
+    ).map(token => {
+      const level = token.tag;
+      const rawText = tokens[tokens.indexOf(token) + 1].content;
+      const text = stripHtml(rawText).result; // Strip HTML tags from the heading text
+      const id = eleventySlugify(text, { lower: true, strict: true, locale: 'fr' });
+      return { level, text, id };
+    });
+
+    // Create TOC HTML
+    let tocHTML = '<aside><h2>{{ onThisPage[locale].heading }}</h2><ul>';
+    const levelsStack = [];
+
+    headings.forEach(heading => {
+      const levelIndex = parseInt(heading.level.substring(1)) - 1;
+
+      while (levelsStack.length && levelsStack[levelsStack.length - 1] > levelIndex) {
+        tocHTML += '</ul></li>';
+        levelsStack.pop();
+      }
+
+      if (levelsStack.length && levelsStack[levelsStack.length - 1] === levelIndex) {
+        tocHTML += '</li>';
+      }
+
+      if (!levelsStack.length || levelsStack[levelsStack.length - 1] < levelIndex) {
+        tocHTML += '<ul>';
+        levelsStack.push(levelIndex);
+      }
+
+      tocHTML += `<li><a href="#${heading.id}">${heading.text}</a>`;
+    });
+
+    while (levelsStack.length) {
+      tocHTML += '</ul></li>';
+      levelsStack.pop();
+    }
+
+    tocHTML += '</ul></aside>';
+
+    return tocHTML;
+  });
+
+  // Return the markdown instance for use by other modules
+  return md;
+};
